LearningContainers/Web scrapers/ALICIA-twitterScraper/requirements.txt
selenium
pandas
python-dotenv
twython
Pillow
flask
Flask-Cors